Clean up folders and files and restructure.
